Regen rabbitmqadmin if version is not the current
Add a custom fact based on "rabbitmqctl version", which is the current
version of the installed rabbitmq-server.

Compare the version that rabbitmqctl and rabbitmqadmin report and if
they are not the same, delete the rabbitrmqadmin file, and thus puppet
will redownload it from the rabbit server, making sure we have the
same version for rabbitmqadmin as the rabbitmq-server.

Notice that rabbitmqctl version is only available on rabbitmq server
3.7.12 and above[1].

[1]https://github.com/rabbitmq/rabbitmq-cli/commit/e5bbc17f7d9d5bb529bc5c20c2776c9a9afe61d1

Change-Id: I78c7750fdf282dea376bab771f2e68f2fa5223bb